Add useInlineSVG hook to handle the lifecycle
