+++ src/modules/hooks.tsx
@@ -1,4 +1,9 @@
-import { useEffect, useRef } from 'react';
+import { EffectCallback, useEffect, useRef } from 'react';
+
+export function useMount(effect: EffectCallback) {
+  // eslint-disable-next-line react-hooks/exhaustive-deps
+  useEffect(effect, []);
+}
 
 export function usePrevious<T>(state: T): T | undefined {
   const ref = useRef<T>(undefined);
